pyproject.toml
[project]
name = "sherlock-ohms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.4.8",
    "duckdb>=1.4.4",
    "google-genai>=1.62.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.9",
    "pikepdf>=10.3.0",
    "plotly>=6.5.2",
    "pyarrow>=18.1.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
lint = "ruff check ."
format = "ruff format ."
test = "pytest ."
run = "python main.py"
pre_test = "task lint"
